# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.InstagramLead:
    properties:
      created_at:
        type: string
      first_interaction:
        type: string
      id:
        type: string
      interaction_count:
        type: integer
      last_interaction:
        type: string
      org_id:
        type: string
      rules_triggered:
        description: rule names
        items:
          type: string
        type: array
      sender_ig_id:
        type: string
      sender_username:
        type: string
      sources:
        description: '"comment", "dm"'
        items:
          type: string
        type: array
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
...
  models.LeadListResponse:
    properties:
      leads:
        items:
          $ref: '#/definitions/models.InstagramLead'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  models.LeadStatsResponse:
    properties:
      by_source:
        additionalProperties:
          format: int64
          type: integer
        type: object
      new_this_week:
        type: integer
      total:
        type: integer
    type: object
...
  models.UpdateLeadTagsRequest:
    properties:
      tags:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8088
info:
  contact: {}
  description: API de autenticacao e gerenciamento de usuarios
  title: Tron Legacy API
  version: "1.0"
paths:
...
  /admin/instagram/leads:
    get:
      description: Retorna lista paginada e filtrável de leads capturados
      parameters:
      - description: Página (padrão 1)
        in: query
        name: page
        type: integer
      - description: Itens por página (padrão 20, máx 100)
        in: query
        name: limit
        type: integer
      - description: Buscar por username
        in: query
        name: search
        type: string
      - description: Filtrar por tag
        in: query
        name: tag
        type: string
      - description: Filtrar por fonte
        in: query
        name: source
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LeadListResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Erro ao buscar leads
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Listar leads do Instagram
      tags:
      - instagram-leads
  /admin/instagram/leads/{id}/tags:
    put:
      consumes:
      - application/json
      description: Atualiza as tags de um lead específico
      parameters:
      - description: ID do lead
        in: path
        name: id
        required: true
        type: string
      - description: Tags para atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateLeadTagsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID ou JSON inválido
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Lead não encontrado
          schema:
            type: string
        "500":
          description: Erro ao atualizar tags
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Atualizar tags de um lead
      tags:
      - instagram-leads
  /admin/instagram/leads/export:
    get:
      description: Exporta todos os leads da organização em formato CSV
      produces:
      - text/csv
      responses:
        "200":
          description: Arquivo CSV
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Erro ao buscar leads
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Exportar leads em CSV
      tags:
      - instagram-leads
  /admin/instagram/leads/stats:
    get:
      description: Retorna estatísticas resumidas dos leads (total, novos na semana,
        por fonte)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LeadStatsResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Erro ao contar leads
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Obter estatísticas de leads
      tags:
      - instagram-leads
...
securityDefinitions:
  BearerAuth:
    description: 'Digite: Bearer {seu_token_aqui}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"